Conflicts:
	Source/WebCore/GNUmakefile.am
	Source/WebCore/GNUmakefile.list.am
	Source/WebKit/gtk/webkit/webkitwebview.cpp
	Source/WebKit2/ChangeLog

+++ Source/WebCore/platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp
@@ -43,6 +43,8 @@
 #include "cc/CCPrioritizedTextureManager.h"
 #include "cc/CCRenderPassDrawQuad.h"
 #include "cc/CCRenderingStats.h"
+#include "cc/CCScrollbarAnimationController.h"
+#include "cc/CCScrollbarLayerImpl.h"
 #include "cc/CCSettings.h"
 #include "cc/CCSingleThreadProxy.h"
 #include <wtf/CurrentTime.h>
@@ -120,7 +122,7 @@ CCLayerTreeHostImpl::CCLayerTreeHostImpl(const CCLayerTreeSettings& settings, CC
     , m_settings(settings)
     , m_deviceScaleFactor(1)
     , m_visible(true)
-    , m_contentsTexturesWerePurgedSinceLastCommit(false)
+    , m_contentsTexturesPurged(false)
     , m_memoryAllocationLimitBytes(CCPrioritizedTextureManager::defaultMemoryAllocationLimit())
     , m_pageScale(1)
     , m_pageScaleDelta(1)
@@ -157,7 +159,6 @@ void CCLayerTreeHostImpl::commitComplete()
     // Recompute max scroll position; must be after layer content bounds are
     // updated.
     updateMaxScrollPosition();
-    m_contentsTexturesWerePurgedSinceLastCommit = false;
 }
 
 bool CCLayerTreeHostImpl::canDraw()
@@ -166,7 +167,7 @@ bool CCLayerTreeHostImpl::canDraw()
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw no root layer");
         return false;
     }
-    if (viewportSize().isEmpty()) {
+    if (deviceViewportSize().isEmpty()) {
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw empty viewport");
         return false;
     }
@@ -174,7 +175,7 @@ bool CCLayerTreeHostImpl::canDraw()
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw no layerRenderer");
         return false;
     }
-    if (m_contentsTexturesWerePurgedSinceLastCommit) {
+    if (m_contentsTexturesPurged) {
         TRACE_EVENT_INSTANT0("cc", "CCLayerTreeHostImpl::canDraw contents textures purged");
         return false;
     }
@@ -191,6 +192,7 @@ void CCLayerTreeHostImpl::animate(double monotonicTime, double wallClockTime)
     animatePageScale(monotonicTime);
     animateLayers(monotonicTime, wallClockTime);
     animateGestures(monotonicTime);
+    animateScrollbars(monotonicTime);
 }
 
 void CCLayerTreeHostImpl::startPageScaleAnimation(const IntSize& targetPosition, bool anchorPoint, float pageScale, double startTime, double duration)
@@ -204,7 +206,7 @@ void CCLayerTreeHostImpl::startPageScaleAnimation(const IntSize& targetPosition,
     IntSize scaledContentSize = contentSize();
     scaledContentSize.scale(m_pageScaleDelta);
 
-    m_pageScaleAnimation = CCPageScaleAnimation::create(scrollTotal, scaleTotal, m_viewportSize, scaledContentSize, startTime);
+    m_pageScaleAnimation = CCPageScaleAnimation::create(scrollTotal, scaleTotal, m_deviceViewportSize, scaledContentSize, startTime);
 
     if (anchorPoint) {
         IntSize windowAnchor(targetPosition);
@@ -253,15 +255,9 @@ void CCLayerTreeHostImpl::calculateRenderSurfaceLayerList(CCLayerList& renderSur
     {
         TRACE_EVENT0("cc", "CCLayerTreeHostImpl::calcDrawEtc");
         CCLayerTreeHostCommon::calculateDrawTransforms(m_rootLayerImpl.get(), deviceViewportSize(), m_deviceScaleFactor, &m_layerSorter, layerRendererCapabilities().maxTextureSize, renderSurfaceLayerList);
+        CCLayerTreeHostCommon::calculateVisibleRects(renderSurfaceLayerList);
 
         trackDamageForAllSurfaces(m_rootLayerImpl.get(), renderSurfaceLayerList);
-
-        if (layerRendererCapabilities().usingPartialSwap)
-            m_rootScissorRect = m_rootLayerImpl->renderSurface()->damageTracker()->currentDamageRect();
-        else
-            m_rootScissorRect = FloatRect(FloatPoint(0, 0), deviceViewportSize());
-
-        CCLayerTreeHostCommon::calculateVisibleAndScissorRects(renderSurfaceLayerList, m_rootScissorRect);
     }
 }
 
@@ -273,8 +269,6 @@ bool CCLayerTreeHostImpl::calculateRenderPasses(FrameData& frame)
 
     TRACE_EVENT1("cc", "CCLayerTreeHostImpl::calculateRenderPasses", "renderSurfaceLayerList.size()", static_cast<long long unsigned>(frame.renderSurfaceLayerList->size()));
 
-    m_rootLayerImpl->setScissorRect(enclosingIntRect(m_rootScissorRect));
-
     // Create the render passes in dependency order.
     HashMap<CCRenderSurface*, CCRenderPass*> surfacePassMap;
     for (int surfaceIndex = frame.renderSurfaceLayerList->size() - 1; surfaceIndex >= 0 ; --surfaceIndex) {
@@ -283,6 +277,7 @@ bool CCLayerTreeHostImpl::calculateRenderPasses(FrameData& frame)
 
         int renderPassId = renderSurfaceLayer->id();
         OwnPtr<CCRenderPass> pass = CCRenderPass::create(renderSurface, renderPassId);
+        pass->setDamageRect(renderSurface->damageTracker()->currentDamageRect());
         pass->setFilters(renderSurfaceLayer->filters());
         pass->setBackgroundFilters(renderSurfaceLayer->backgroundFilters());
 
@@ -292,7 +287,7 @@ bool CCLayerTreeHostImpl::calculateRenderPasses(FrameData& frame)
     }
 
     bool recordMetricsForFrame = true; // FIXME: In the future, disable this when about:tracing is off.
-    CCOcclusionTrackerImpl occlusionTracker(enclosingIntRect(m_rootScissorRect), recordMetricsForFrame);
+    CCOcclusionTrackerImpl occlusionTracker(m_rootLayerImpl->renderSurface()->contentRect(), recordMetricsForFrame);
     occlusionTracker.setMinimumTrackingSize(m_settings.minimumOcclusionTrackingSize);
 
     if (settings().showOccludingRects)
@@ -314,10 +309,10 @@ bool CCLayerTreeHostImpl::calculateRenderPasses(FrameData& frame)
 
         occlusionTracker.enterLayer(it);
 
-        if (it.representsContributingRenderSurface() && !it->renderSurface()->scissorRect().isEmpty()) {
+        if (it.representsContributingRenderSurface()) {
             CCRenderPass* contributingRenderPass = surfacePassMap.get(it->renderSurface());
             pass->appendQuadsForRenderSurfaceLayer(*it, contributingRenderPass, &occlusionTracker);
-        } else if (it.representsItself() && !it->visibleContentRect().isEmpty() && !it->scissorRect().isEmpty()) {
+        } else if (it.representsItself() && !it->visibleContentRect().isEmpty()) {
             bool hasOcclusionFromOutsideTargetSurface;
             if (occlusionTracker.occluded(*it, it->visibleContentRect(), &hasOcclusionFromOutsideTargetSurface)) {
                 if (hasOcclusionFromOutsideTargetSurface)
@@ -516,8 +511,10 @@ bool CCLayerTreeHostImpl::prepareToDraw(FrameData& frame)
 
 void CCLayerTreeHostImpl::releaseContentsTextures()
 {
+    if (m_contentsTexturesPurged)
+        return;
     m_resourceProvider->deleteOwnedResources(CCRenderer::ContentPool);
-    m_contentsTexturesWerePurgedSinceLastCommit = true;
+    m_contentsTexturesPurged = true;
     m_client->setNeedsCommitOnImplThread();
 }
 
@@ -531,6 +528,11 @@ void CCLayerTreeHostImpl::setMemoryAllocationLimitBytes(size_t bytes)
     m_client->setNeedsCommitOnImplThread();
 }
 
+void CCLayerTreeHostImpl::onVSyncParametersChanged(double monotonicTimebase, double intervalInSeconds)
+{
+    m_client->onVSyncParametersChanged(monotonicTimebase, intervalInSeconds);
+}
+
 void CCLayerTreeHostImpl::drawLayers(const FrameData& frame)
 {
     TRACE_EVENT0("cc", "CCLayerTreeHostImpl::drawLayers");
@@ -542,15 +544,19 @@ void CCLayerTreeHostImpl::drawLayers(const FrameData& frame)
     // RenderWidget.
     m_fpsCounter->markBeginningOfFrame(currentTime());
 
-    m_layerRenderer->drawFrame(frame.renderPasses, frame.renderPassesById, m_rootScissorRect);
-
-    for (unsigned int i = 0; i < frame.renderPasses.size(); i++)
-        frame.renderPasses[i]->targetSurface()->damageTracker()->didDrawDamagedArea();
+    m_layerRenderer->drawFrame(frame.renderPasses, frame.renderPassesById);
 
     if (m_settings.showDebugRects())
         m_debugRectHistory->saveDebugRectsForCurrentFrame(m_rootLayerImpl.get(), *frame.renderSurfaceLayerList, frame.occludingScreenSpaceRects, settings());
 
+    // Once a RenderPass has been drawn, its damage should be cleared in
+    // case the RenderPass will be reused next frame.
+    for (unsigned int i = 0; i < frame.renderPasses.size(); i++)
+        frame.renderPasses[i]->setDamageRect(FloatRect());
+
     // The next frame should start by assuming nothing has changed, and changes are noted as they occur.
+    for (unsigned int i = 0; i < frame.renderSurfaceLayerList->size(); i++)
+        (*frame.renderSurfaceLayerList)[i]->renderSurface()->damageTracker()->didDrawDamagedArea();
     m_rootLayerImpl->resetAllChangeTrackingForSubtree();
 }
 
@@ -581,7 +587,7 @@ bool CCLayerTreeHostImpl::swapBuffers()
     ASSERT(m_layerRenderer);
 
     m_fpsCounter->markEndOfFrame();
-    return m_layerRenderer->swapBuffers(enclosingIntRect(m_rootScissorRect));
+    return m_layerRenderer->swapBuffers();
 }
 
 void CCLayerTreeHostImpl::didLoseContext()
@@ -675,7 +681,11 @@ void CCLayerTreeHostImpl::setVisible(bool visible)
 
 bool CCLayerTreeHostImpl::initializeLayerRenderer(PassOwnPtr<CCGraphicsContext> context, TextureUploaderOption textureUploader)
 {
+    if (!context->bindToClient(this))
+        return false;
+
     WebKit::WebGraphicsContext3D* context3d = context->context3D();
+
     if (!context3d) {
         // FIXME: Implement this path for software compositing.
         return false;
@@ -705,15 +715,13 @@ bool CCLayerTreeHostImpl::initializeLayerRenderer(PassOwnPtr<CCGraphicsContext> 
     return m_layerRenderer;
 }
 
-void CCLayerTreeHostImpl::setViewportSize(const IntSize& viewportSize)
+void CCLayerTreeHostImpl::setViewportSize(const IntSize& layoutViewportSize, const IntSize& deviceViewportSize)
 {
-    if (viewportSize == m_viewportSize)
+    if (layoutViewportSize == m_layoutViewportSize && deviceViewportSize == m_deviceViewportSize)
         return;
 
-    m_viewportSize = viewportSize;
-
-    m_deviceViewportSize = viewportSize;
-    m_deviceViewportSize.scale(m_deviceScaleFactor);
+    m_layoutViewportSize = layoutViewportSize;
+    m_deviceViewportSize = deviceViewportSize;
 
     updateMaxScrollPosition();
 
@@ -742,12 +750,6 @@ void CCLayerTreeHostImpl::setDeviceScaleFactor(float deviceScaleFactor)
     if (deviceScaleFactor == m_deviceScaleFactor)
         return;
     m_deviceScaleFactor = deviceScaleFactor;
-
-    m_deviceViewportSize = viewportSize();
-    m_deviceViewportSize.scale(m_deviceScaleFactor);
-    updateMaxScrollPosition();
-    if (m_layerRenderer)
-        m_layerRenderer->viewportChanged();
 }
 
 
@@ -799,13 +801,15 @@ void CCLayerTreeHostImpl::updateMaxScrollPosition()
     if (!m_rootScrollLayerImpl || !m_rootScrollLayerImpl->children().size())
         return;
 
-    FloatSize viewBounds = m_viewportSize;
+    FloatSize viewBounds = m_deviceViewportSize;
     if (CCLayerImpl* clipLayer = m_rootScrollLayerImpl->parent()) {
-        if (clipLayer->masksToBounds())
+        // Compensate for non-overlay scrollbars.
+        if (clipLayer->masksToBounds()) {
             viewBounds = clipLayer->bounds();
+            viewBounds.scale(m_deviceScaleFactor);
+        }
     }
     viewBounds.scale(1 / m_pageScaleDelta);
-    viewBounds.scale(m_deviceScaleFactor);
 
     // maxScroll is computed in physical pixels, but scroll positions are in layout pixels.
     IntSize maxScroll = contentSize() - expandedIntSize(viewBounds);
@@ -931,6 +935,9 @@ void CCLayerTreeHostImpl::pinchGestureBegin()
 {
     m_pinchGestureActive = true;
     m_previousPinchAnchor = IntPoint();
+
+    if (m_rootScrollLayerImpl && m_rootScrollLayerImpl->scrollbarAnimationController())
+        m_rootScrollLayerImpl->scrollbarAnimationController()->didPinchGestureBegin();
 }
 
 void CCLayerTreeHostImpl::pinchGestureUpdate(float magnifyDelta,
@@ -954,6 +961,10 @@ void CCLayerTreeHostImpl::pinchGestureUpdate(float magnifyDelta,
     m_previousPinchAnchor = anchor;
 
     m_rootScrollLayerImpl->scrollBy(roundedIntSize(move));
+
+    if (m_rootScrollLayerImpl->scrollbarAnimationController())
+        m_rootScrollLayerImpl->scrollbarAnimationController()->didPinchGestureUpdate();
+
     m_client->setNeedsCommitOnImplThread();
     m_client->setNeedsRedrawOnImplThread();
 }
@@ -961,6 +972,9 @@ void CCLayerTreeHostImpl::pinchGestureUpdate(float magnifyDelta,
 void CCLayerTreeHostImpl::pinchGestureEnd()
 {
     m_pinchGestureActive = false;
+
+    if (m_rootScrollLayerImpl && m_rootScrollLayerImpl->scrollbarAnimationController())
+        m_rootScrollLayerImpl->scrollbarAnimationController()->didPinchGestureEnd();
 
     m_client->setNeedsCommitOnImplThread();
 }
@@ -1177,6 +1191,24 @@ void CCLayerTreeHostImpl::renderingStats(CCRenderingStats& stats) const
 {
     stats.numFramesSentToScreen = fpsCounter()->currentFrameNumber();
     stats.droppedFrameCount = fpsCounter()->droppedFrameCount();
+}
+
+void CCLayerTreeHostImpl::animateScrollbars(double monotonicTime)
+{
+    animateScrollbarsRecursive(m_rootLayerImpl.get(), monotonicTime);
+}
+
+void CCLayerTreeHostImpl::animateScrollbarsRecursive(CCLayerImpl* layer, double monotonicTime)
+{
+    if (!layer)
+        return;
+
+    CCScrollbarAnimationController* scrollbarController = layer->scrollbarAnimationController();
+    if (scrollbarController && scrollbarController->animate(monotonicTime))
+        m_client->setNeedsRedrawOnImplThread();
+
+    for (size_t i = 0; i < layer->children().size(); ++i)
+        animateScrollbarsRecursive(layer->children()[i].get(), monotonicTime);
 }
 
 } // namespace WebCore
